Clean up Excel export
